x86 8-bit immediate values were interpreted incorrectly cross-platform.
Correct interpretation of Intel instructions requires
sign extension.   On intel it works because chars are
signed on that platform.   On other platforms (arm64, powerpc64)
intel code was not correctly analyzed because incorrect
values were generated from 8 bit immediate constants;
which were interpreted as unsigned on those systems.

These changes ensure intel code will be correctly interpreted
on any system, regardless of the signedness of chars.

A n-way validation was used to validate this change, to
ensure that correct code interpretation of (intel, arm,
power) code happens on all three platforms.    No other
issues of this nature were located -- at this time.

Jim Kupsch contributed additional correctness changes to
common/src/arch-x86.C that were added to this
branch/PR to fix yet more problems of this ilk.

+++ common/src/arch-x86.C
@@ -10490,7 +10490,7 @@ static unsigned int ia32_decode_modrm(const unsigned int addrSzAttr, const unsig
    addr++;
 
    /* Get displacements we're going to use */
-   const char* disp8 = (const char*)addr;
+   const signed char* disp8 = (const signed char*)addr;
    const short* disp16 = (const short*)addr;
    const int* disp32 = (const int*)addr;
 
@@ -10675,7 +10675,7 @@ static unsigned int ia32_decode_modrm(const unsigned int addrSzAttr, const unsig
       }
 
       /* Update displacement pointers  */
-      disp8 = (const char*)addr;
+      disp8 = (const signed char*)addr;
       disp16 = (const short*)addr;
       disp32 = (const int*)addr;
 
@@ -10953,7 +10953,7 @@ unsigned int ia32_decode_operands (const ia32_prefixes& pref,
                nib += wordSzB * addrSzAttr;
                if(mac)
                {
-                  int offset = 0;
+                  long offset = 0;
                   switch(addrSzAttr)
                   {
                      case 1: // 16-bit offset
@@ -11744,7 +11744,7 @@ int displacement(const unsigned char *instr, unsigned type) {
       disp = *(const int *)(instr+2);
    } else if (type & IS_JUMP) {
       if (type & REL_B) {
-         disp = *(const char *)(instr+1);
+         disp = *(const signed char *)(instr+1);
       } else if (type & REL_W) {
          disp = *(const short *)(instr+1); // skip opcode
       } else if (type & REL_D) {
@@ -11752,7 +11752,7 @@ int displacement(const unsigned char *instr, unsigned type) {
       }
    } else if (type & IS_JCC) {
       if (type & REL_B) {
-         disp = *(const char *)(instr+1);
+         disp = *(const signed char *)(instr+1);
       } else if (type & REL_W) {
          disp = *(const short *)(instr+2); // skip two byte opcode
       } else if (type & REL_D) {
